Create gateway utils for securing routes
